# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentação da API do sistema",
        "contact": {
            "email": "contato@exemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8989"
        }
    ],
    "paths": {
...
        "/api/product-price/{id}": {
            "get": {
                "tags": [
                    "Product Prices"
                ],
                "summary": "Obtém um preço de produto específico",
                "description": "Retorna os detalhes de um preço de produto específico pelo ID",
                "operationId": "getProductPrice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do preço do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do preço do produto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductPrice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Preço do produto não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product Prices"
                ],
                "summary": "Atualiza um preço de produto existente",
                "description": "Atualiza os dados de um preço de produto específico",
                "operationId": "updateProductPrice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do preço do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductPriceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preço do produto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductPrice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Preço do produto não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product Prices"
                ],
                "summary": "Remove um preço de produto",
                "description": "Remove um preço de produto específico do sistema",
                "operationId": "destroyProductPrice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do preço do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Preço do produto removido com sucesso"
                    },
                    "404": {
                        "description": "Preço do produto não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product-price": {
            "get": {
                "tags": [
                    "Product Prices"
                ],
                "summary": "Lista todos os preços de produtos",
                "description": "Retorna uma lista de todos os preços de produtos cadastrados",
                "operationId": "indexProductPrice",
                "responses": {
                    "200": {
                        "description": "Lista de preços de produtos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductPrice"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product Prices"
                ],
                "summary": "Cria um novo preço de produto",
                "description": "Cria um novo preço de produto com os dados fornecidos",
                "operationId": "storeProductPrice",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductPriceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Preço de produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductPrice"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/promotion/{id}": {
            "get": {
                "tags": [
                    "Promotions"
                ],
                "summary": "Obtém uma promoção específica",
                "description": "Retorna os detalhes de uma promoção específica pelo ID",
                "operationId": "getPromotion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da promoção",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da promoção",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Promotion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Promoção não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Promotions"
                ],
                "summary": "Atualiza uma promoção existente",
                "description": "Atualiza os dados de uma promoção específica",
                "operationId": "updatePromotion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da promoção",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePromotionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Promoção atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Promotion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Promoção não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Promotions"
                ],
                "summary": "Remove uma promoção",
                "description": "Remove uma promoção específica do sistema",
                "operationId": "destroyPromotion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da promoção",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Promoção removida com sucesso"
                    },
                    "404": {
                        "description": "Promoção não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Promotion": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Promoção de Verão"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descontos especiais para o verão"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "fixed_discount",
                            "percentage_discount",
                            "buy_x_get_y"
                        ],
                        "example": "fixed_discount"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-31T23:59:59Z"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "UpdatePromotionRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Promoção de Verão"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descontos especiais para o verão"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "fixed_discount",
                            "percentage_discount",
                            "buy_x_get_y"
                        ],
                        "example": "fixed_discount"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-31T23:59:59Z"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "ProductPrice": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "currency": {
                        "type": "string",
                        "example": "BRL"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-31T23:59:59Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "UpdateProductPriceRequest": {
                "properties": {
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "currency": {
                        "type": "string",
                        "example": "BRL"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-31T23:59:59Z"
                    }
                },
                "type": "object"
            },
...
            "StoreProductPriceRequest": {
                "required": [
                    "product_id",
                    "price",
                    "start_date"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1999.99
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-19"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-19",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}